+++ src/include/opcodes_ppc.h
@@ -252,6 +252,8 @@
 #define	  PPC_63_FMUL		  25
 #define	  PPC_63_FMSUB		  28
 #define	  PPC_63_FMADD		  29
+#define	  PPC_63_FNMSUB		  30	/*  #344  */
+#define	  PPC_63_FNMADD		  31	/*  #344  */
 #define	  PPC_63_FNEG		  40
 #define	  PPC_63_FMR		  72
 #define	  PPC_63_FNABS		  136